completed file structure reorganisation
