+++ src/utils/utils.js
@@ -1,7 +1,7 @@
-import { getCLEArtDetailsByID, getMETAllArtworkIDs, getMETArtDetails } from "./api";
-import noImagePicture from './assets/img/No-Image-Placeholder.svg'
-import { db } from './config/firebase';
-import { doc, getDoc, updateDoc, setDoc, arrayRemove, arrayUnion } from 'firebase/firestore'
+import { getCLEArtDetailsByID, getMETAllArtworkIDs, getMETArtDetails } from "../config/api.js";
+import noImagePicture from '../assets/img/No-Image-Placeholder.svg'
+import { db } from '../config/firebase';
+import { doc, getDoc, updateDoc, setDoc } from 'firebase/firestore'
 
 export const RandomArtID = async () => {
   const allIDsResponse = await getMETAllArtworkIDs();
